% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_p.R
\name{correct_p}
\alias{correct_p}
\title{Take a list containing model/test objects and return them with jointly corrected p-values}
\usage{
correct_p(objs)
}
\arguments{
\item{objs}{A list of objects to correct}
}
\value{
A list of objects with updated p values
}
\description{
Take a list containing model/test objects and return them with jointly corrected p-values
}
