% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customR-package.R
\docType{package}
\name{customR-pkg}
\alias{customR-pkg}
\alias{customR-pkg-package}
\title{\pkg{customR}: Custom utility functions}
\description{
customR contains custom functions that were developed from stackoverflow and reddit users that had a need for certain functions
}
\note{
Development of this package was kindly supported by
CK under CK-GOFUNDME
}
\section{customR functions}{

The main functions are:
\tabular{ll}{
  \code{rerank()}        \tab Remove and Rank\cr
  }
}

\references{
Chong H. Kim et al. (2019). \emph{Development of random/custom
functions for users of stackoverflow and reddit r/rstats}
In preparation.
}
\seealso{
\code{PMMSKNN}
}
