% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printp.R
\name{printp}
\alias{printp}
\title{Return string with formatted p-value}
\usage{
printp(val)
}
\arguments{
\item{val}{A p-value to format.}
}
\value{
A string formatted in APA style with the =/< sign and the p-value
}
\description{
Return string with formatted p-value
}
