% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerank.R
\name{rerank}
\alias{rerank}
\title{Remove and Rank observations in a dataframe}
\usage{
rerank(df, var, thresh, upper = TRUE, inclusive = TRUE,
  ascending = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{var}{Variable name. Can be quoted or unquoted}

\item{thresh}{Numeric value indicating number to use to
filter the data}

\item{upper}{Logical (TRUE/FALSE) indicating wheter to
filter values above the threshold. Default is `TRUE`}

\item{inclusive}{Logical (TRUE/FALSE) indicating wheter to
Include the threshold value in filtering}

\item{ascending}{Logical (TRUE/FALSE) indicating wheter to
Rank the rows by the `var` in ascending sequence}
}
\value{
dataframe that is filtered and ranked
}
\description{
`rerank()` takes a dataframe, a column name (numeric)
and a threshold value for filtering and a string that
specifies whether to remove values above or below the
threshold value
}
